refactor: add missing schemas and renamed files
