+++ scrapegraphai/graphs/pdf_scraper_multi_graph.py
